move all home page components to it own folder
