+++ app/javascript/components/Home/PageIndicator.js
