+++ app/javascript/components/Session.js
@@ -4,7 +4,7 @@ import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import SignUp from './SignUp';
 import Login from './Login';
 import Navbar from './Navbar';
-import HomePage from './HomePage';
+import HomePage from './Home/HomePage';
 
 const Session = () => {
   const currentSession = useSelector((state) => state.session);
